requirements.txt
request
pillow